Split schema detection out of step_text.rs, which went over the ratchet
The Rust module-size ratchet failed: rust/export/src/step_text.rs reached 420
lines against a 400 limit with no allowlist row. Same house rule as the
TypeScript side, split rather than allowlist.

`find_unquoted` and `detect_schema` move to `schema_detect.rs`, and their ten
tests move with them to `schema_detect_tests.rs`. The seam is real rather than
convenient: `step_text.rs` is about EDITING STEP text, and this is about
reading one fact out of it before anything is parsed. They also have different
hazards, which is most of why the bugs this branch fixed lived here. Everything
in the new module runs on whole uncapped attacker-supplied files, before any
structure is known to exist, and the module doc says so.

260 and 180 lines. No logic changed; the 336 export tests pass identically.

Note on how this was found: `cargo test --workspace` was already red when the
previous commit was made. It was run, it reported 101, and the commit went
ahead anyway because it was chained behind the test in one command line. The
gate did its job; the shell hid it. Same shape as `cmd | tail; echo $?`.

Verified by exit code, each run on its own: export tests 0, cargo test
--workspace 0, clippy -D warnings 0, module-size 0, typecheck 0, lint 0.
