+++ rust/export/src/step.rs
@@ -112,8 +112,9 @@ pub struct StepStats {
     pub copies_refused: usize,
 }
 
+use crate::schema_detect::detect_schema;
 use crate::step_text::{
-    apply_attr_mutations, detect_schema, escape, merge_edits, refs_in_line, renumber,
+    apply_attr_mutations, escape, merge_edits, refs_in_line, renumber,
 };
 
 /// Export the parsed model in `content` as a STEP/IFC string.
